+++ CHANGELOG.md
@@ -9,6 +9,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Added ⭐
+
+* `egui::popup::popup_below_widget`: show a popup area below another widget
+
 
 ## 0.8.0 - 2021-01-17 - Grid layout & new visual style